+++ src/util/factories/ProdutoFactory.java
@@ -4,7 +4,7 @@ import model.Categoria;
 import model.Produto;
 
 public class ProdutoFactory {
-    public static Produto criarProduto(String nome,String descricao, double preco, int quantidade, Categoria categoria) {
+    public static Produto criarProduto(String nome,String descricao, float preco, int quantidade, Categoria categoria) {
         return new Produto(nome, descricao, preco, quantidade, categoria);
     }
 }
